feat(response)!: keyed/non-keyed responses
Changes
-
- Rename the "Response" functions to "KeyedResponse" functions
to align with conventions within express-wrapper. These functions
use string-based "type" values alongside the payload.
- Add "Response" functions in place of the original, further
aligning with conventions within express-wrapper.  These functions
use numerical-based "type" values alongside the payload.
- Add "conflict" / 409 error code to the possible status that can be
returned

BREAKING CHANGE: Existing response functions have been renamed to
align convention with other packages in this repository.
